Automation_Selenium/Browser_Multiple/requirements.txt
selenium
webdriver-manager
PyYAML
pytest
